+++ raffia/src/error.rs
@@ -96,6 +96,7 @@ pub enum ErrorKind {
     UnexpectedLessMixinCall,
     UnexpectedSemicolonInSass,
     UnexpectedSimpleBlock,
+    TopLevelDeclaration,
 }
 
 impl Display for ErrorKind {
@@ -215,6 +216,7 @@ impl Display for ErrorKind {
             Self::UnexpectedLessMixinCall => write!(f, "Less mixin call is disallowed"),
             Self::UnexpectedSemicolonInSass => write!(f, "semicolon in Sass is disallowed"),
             Self::UnexpectedSimpleBlock => write!(f, "simple block is disallowed"),
+            Self::TopLevelDeclaration => write!(f, "declaration at top level is disallowed"),
         }
     }
 }